+++ packages/react/src/index.ts
@@ -75,3 +75,9 @@ export { Slider } from "./components/Slider/Slider";
 export type { SliderProps } from "./components/Slider/Slider";
 export { NumberInput } from "./components/NumberInput/NumberInput";
 export type { NumberInputProps } from "./components/NumberInput/NumberInput";
+export { SearchInput } from "./components/SearchInput/SearchInput";
+export type {
+  SearchInputProps,
+  SearchInputSuggestions,
+  SearchInputSuggestionObject,
+} from "./components/SearchInput/SearchInput";